pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "python-telegram-bot==20.7",
    "requests>=2.32.4",
    "werkzeug>=3.1.3",
]

[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["lua lia <45708543-hyejooteamo@users.noreply.replit.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
